+++ backend/src/main/java/com/repository/ProductRepository.java
@@ -2,6 +2,7 @@ package com.repository;
 
 import com.entity.Category;
 import com.entity.Product;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,13 +10,13 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
-    public Optional<Product> findByProductId(Long productId);
+    public Optional<Product> findByProductId(Integer productId);
 
     public Optional<Product> findByProductName(String productName);
 
-    public List<Product> findByCategoryId(Long categoryId);
+    public List<Product> findByCategoryId(Integer categoryId);
 
     List<Product> findByProductNameContainingIgnoreCase(String productName);
 }
